Split middleware tests into separate files
