+++ src/store/middleware/setQueriesFromStateMiddleware.test.ts
@@ -1,5 +1,3 @@
-import addMetaToRoutesMiddleware from './addMetaToRoutesMiddleware'
-import preserveUrlParametersMiddleware from './preserveUrlParametersMiddleware'
 import setQueriesFromStateMiddleware from './setQueriesFromStateMiddleware'
 import paramsRegistry from '../params-registry'
 
@@ -13,7 +11,6 @@ Object.defineProperties(global, {
 describe('Custom Redux Middleware', () => {
   const isRouterTypeMock = jest.spyOn(paramsRegistry, 'isRouterType')
   const setQueriesFromStateMock = jest.spyOn(paramsRegistry, 'setQueriesFromState')
-  const getParametersForRouteMock = jest.spyOn(paramsRegistry, 'getParametersForRoute')
   const mockStore = {
     getState: jest.fn(() => ({
       location: {
@@ -22,8 +19,6 @@ describe('Custom Redux Middleware', () => {
     })),
   }
 
-  const nextMockAddMetaToRoutes = jest.fn()
-  const nextMockPreserveUrlParametersMiddleware = jest.fn()
   const action = { type: 'some action', meta: { query: 'someQuery', preserve: true } }
 
   it('should skip when pathname includes "kaart"', () => {
@@ -32,17 +27,10 @@ describe('Custom Redux Middleware', () => {
       pathname: '/kaart/foo/bar',
     }
 
-    addMetaToRoutesMiddleware(mockStore)(nextMockAddMetaToRoutes)(action)
-    expect(nextMockAddMetaToRoutes).toHaveBeenCalledWith(action)
-
-    preserveUrlParametersMiddleware()(nextMockPreserveUrlParametersMiddleware)(action)
-    expect(nextMockPreserveUrlParametersMiddleware).toHaveBeenCalledWith(action)
-
     setQueriesFromStateMiddleware(mockStore)(jest.fn)(action)
 
     expect(setQueriesFromStateMock).not.toHaveBeenCalled()
     expect(isRouterTypeMock).not.toHaveBeenCalled()
-    expect(getParametersForRouteMock).not.toHaveBeenCalled()
   })
 
   it('should use the custom middleware when pathname does not includes "kaart"', () => {
@@ -52,14 +40,9 @@ describe('Custom Redux Middleware', () => {
       search: '?foo=bar',
     }
 
-    addMetaToRoutesMiddleware(mockStore)(nextMockAddMetaToRoutes)(action)
-
-    preserveUrlParametersMiddleware()(nextMockPreserveUrlParametersMiddleware)(action)
-
     setQueriesFromStateMiddleware(mockStore)(jest.fn)(action)
 
     expect(setQueriesFromStateMock).toHaveBeenCalled()
     expect(isRouterTypeMock).toHaveBeenCalled()
-    expect(getParametersForRouteMock).toHaveBeenCalled()
   })
 })
